MCTS ready

+++ code/State.py
@@ -21,7 +21,8 @@ class State():
             2: 'd', 
             3: 'a', 
             4: 'attack',
-            5: 'cut'
+            5: 'cut',
+            6: 'closer'
         }
         self.data_objects = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'dataset_objects','data.yaml')
         self.data_tiles = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'dataset_tiles', 'data.yaml')
@@ -46,6 +47,8 @@ class State():
 
         self.inventory = Inventory(self.classes)
         self.map = Map(self.classes)
+
+        self.score = 0
     
     def run_action(self, action):
         new_state = State()
@@ -61,34 +64,33 @@ class State():
                 # Move map downward
                 self.map.moveMap(new_state.map.current_map, new_i=1)
                 reward = 1
-        elif action == 2: # d
-            if self.map.canMove(right=True):
-                # Move map left
-                self.map.moveMap(new_state.map.current_map, new_j=-1)
-                reward = 1
             else:
                 new_state.map.current_map = self.map.current_map.copy()
-                reward = 0
+        elif action == 2: # d
+            if self.map.canMove(right=True):
+                # Move map right
+                self.map.moveMap(new_state.map.current_map, new_j=-1)
+                reward = 2
+            else:
+                new_state.map.current_map = self.map.current_map.copy()
         elif action == 3: # a
             if self.map.canMove(right=False):
                 self.map.moveMap(new_state.map.current_map, new_j=1)
                 # Move map left
-                reward = 1
+                reward = 2
             else:
                 new_state.map.current_map = self.map.current_map.copy()
-                reward = 0
         elif action == 4: # attack
             # find closest enemy position and check 
             attack, x, y = self.map.isEnemyOnAttackRange()
             # if is in attack range
             new_state.map.current_map = self.map.current_map.copy()
             if attack:
-                # delete slime tiles from map accroding to x and y
-                new_state.map.deleteTileAt(x, y, self.classes.index('slime'))      
-                reward = 2
-            else:
-                # if not
-                reward = 0
+                # delete tree tiles from map accroding to x and y
+                if(self.last_action == 4):
+                    new_state.map.deleteTileAt(x, y, self.classes.index('slime'))
+                    reward = 2 # to extra to kill a slime
+                reward += 11
         elif action == 5: # cut wood
             # find closest tree position and check 
             # if is in cut range
@@ -96,23 +98,36 @@ class State():
             # if is in attack range
             new_state.map.current_map = self.map.current_map.copy()
             if cut:
-                # delete tree tiles from map accroding to x and y
-                new_state.map.deleteTileAt(x, y, self.classes.index('tree'))
-                self.cut_tree += 1    
-                reward = 4
+                # delete tree tiles, after second cut, from map acording to x and y
+                if(self.last_action == 5):
+                    new_state.map.deleteTileAt(x, y, self.classes.index('tree'))
+                    self.cut_tree += 1    
+                    reward = 2 # to extra to cut the tree
+                reward += 10
+        elif action == 6: # get closer to a tree
+            reward = 4
+            closest = self.map.getCloser()
+            if closest - 58 > 0:
+                right = True
+                #action = 2
             else:
-                # if not
-                reward = 0
-
+                right = False
+                #action = 3
+            if right and self.map.canMove(right):
+                # Move map left
+                self.map.moveMap(new_state.map.current_map, new_j=-1)
+            elif not right and self.map.canMove(right):
+                self.map.moveMap(new_state.map.current_map, new_j=1)
         new_state.last_action = action
         # Apply effects of the map, gravity attack by an enemy
-        new_state.map.fixMap(action == 1)
+        new_state.map.fixMap(action == 1 and self.last_action != 1)
 
         new_state.map.apply_damage()
 
         new_health = new_state.map.getHealth()
         if new_health - health < 0:
-            reward = reward - 2
+            reward = reward - 5
+        new_state.score = self.score + reward
 
         return new_state, reward
 
@@ -132,10 +147,10 @@ class State():
 
         return done
 
-    def get_available_actions(self, state):
-        actions = [0]
-        if self.map.canJump():
-            actions.append(1)
+    def get_available_actions(self):
+        actions = [0, 6]
+        #if self.map.canJump():
+        #    actions.append(1)
         if self.map.canMove(right=True):
             actions.append(2)
         if self.map.canMove(right=False):
@@ -144,7 +159,15 @@ class State():
             actions.append(4)
         if self.map.isTreeOnCutRange()[0]:
             actions.append(5)
-        return actions       
+        return actions
+
+    def __str__(self) -> str:
+        return str(self.map) + str(self.inventory)
+    
+    def print(self):
+        with open("delete.txt", 'w') as f:
+            f.write(str(self.map))
+
 
 
 if __name__ == "__main__":
